Reorganized and cleaned up dircon_utils. -Moved to solvers/ directory as optimization_utils, since it is not dircon specific. -Added documentation. -Made use of new MathematicalProgram.GetAllConstraints(). -Changed all APIs to use better naming schemes and appropriate use of const references and pointers.
