+++ examples/PlanarWalker/run_gait_dircon.cc
@@ -18,7 +18,7 @@
 
 #include "common/find_resource.h"
 #include "systems/primitives/subvector_pass_through.h"
-#include "systems/trajectory_optimization/dircon_util.h"
+#include "solvers/optimization_utils.h"
 #include "systems/trajectory_optimization/dircon_position_data.h"
 #include "systems/trajectory_optimization/dircon_kinematic_data_set.h"
 #include "systems/trajectory_optimization/hybrid_dircon.h"
@@ -233,8 +233,7 @@ shared_ptr<HybridDircon<double>> runDircon(
   MatrixXd A;
   VectorXd y, lb, ub;
   VectorXd x_sol = result.get_x_val();
-  systems::trajectory_optimization::linearizeConstraints(trajopt.get(),
-    x_sol, y, A, lb, ub);
+  solvers::LinearizeConstraints(*trajopt, x_sol, &y, &A, &lb, &ub);
 
 //  MatrixXd y_and_bounds(y.size(),3);
 //  y_and_bounds.col(0) = lb;
